+++ modules/azure-ad/azad-kube-proxy/outputs.tf
@@ -1,7 +1,7 @@
 output "data" {
   description = "The Azure AD Application data for azad-kube-proxy"
   value = {
-    client_id     = azuread_application.this.application_id
+    client_id     = azuread_application.this.client_id
     client_secret = azuread_application_password.this.value
     tenant_id     = data.azuread_client_config.current.tenant_id
   }