+++ modules/azure/aks/main.tf
@@ -17,11 +17,11 @@ terraform {
 
   required_providers {
     azurerm = {
-      version = "3.71.0"
+      version = "3.99.0"
       source  = "hashicorp/azurerm"
     }
     azuread = {
-      version = "2.41.0"
+      version = "2.47.0"
       source  = "hashicorp/azuread"
     }
     random = {