+++ modules/kubernetes/falco/variables.tf
@@ -1,4 +1,4 @@
-variable "cloud_provider" {
-  description = "Cloud provider used for falco"
+variable "cluster_id" {
+  description = "Unique identifier of the cluster across regions and instances."
   type        = string
 }