+++ modules/kubernetes/node-local-dns/main.tf
@@ -25,7 +25,8 @@ resource "git_repository_file" "kustomization" {
 resource "git_repository_file" "node_local_dns" {
   path = "platform/${var.cluster_id}/node-local-dns/node-local-dns.yaml"
   content = templatefile("${path.module}/templates/node-local-dns.yaml.tpl", {
-    local_dns = "169.254.20.10"
-    dns_ip    = var.dns_ip
+    local_dns        = "169.254.20.10"
+    dns_ip           = var.dns_ip
+    coredns_upstream = var.coredns_upstream
   })
 }