+++ validation/azure/aks/main.tf
@@ -18,6 +18,7 @@ module "aks" {
   environment                   = "dev"
   location_short                = "we"
   name                          = "xks"
+  subscription_name             = "sub"
   core_name                     = "core"
   unique_suffix                 = 1337
   aks_name_suffix               = "1"
@@ -56,8 +57,7 @@ module "aks" {
   }
   namespaces = [
     {
-      name                    = "team1"
-      delegate_resource_group = true
+      name = "team1"
       labels = {
         "test" = "test"
       }
@@ -100,4 +100,7 @@ module "aks" {
 
   log_eventhub_name                  = "eventhub1"
   log_eventhub_authorization_rule_id = "12345"
+
+  global_location_short = "we"
+  dns_zones             = ["example.com"]
 }