+++ validation/kubernetes/external-dns/main.tf
@@ -1,17 +1,9 @@
 terraform {}
 
-provider "kubernetes" {}
-
-provider "helm" {}
-
 module "external_dns" {
   source = "../../../modules/kubernetes/external-dns"
 
-  providers = {
-    kubernetes = kubernetes
-    helm       = helm
-  }
-
+  cluster_id   = "foo"
   dns_provider = "azure"
   txt_owner_id = "dev-aks1"
   azure_config = {
@@ -19,6 +11,5 @@ module "external_dns" {
     subscription_id = "id"
     resource_group  = "name"
     client_id       = "id"
-    resource_id     = "id"
   }
 }