+++ validation/kubernetes/falco/main.tf
@@ -7,10 +7,5 @@ provider "helm" {}
 module "falco" {
   source = "../../../modules/kubernetes/falco"
 
-  providers = {
-    kubernetes = kubernetes
-    helm       = helm
-  }
-
-  cloud_provider = "bar"
+  cluster_id = "foo"
 }